Seems have found the bug in "the test runner"
Everywhere the interpreter opens a file, it uses open() with "rb", which means it reads
the '\r' as well. In io_read_file and more places.
We need to have the interpreter work with open() with "r", and this
requires changing how we determine the opened file's size.
Also, if haven't done this already, refactor the interpreter to use the
same io_read_file everywhere, eliminate duplication.