Where were we in description
Now composing new ObjectStrings by not calling object_string_new
recursively for the attributes, since this leads to a stack overflow.

Instead, implemented "directly" cell_table and table functions to put an
ObjectString Value directly in the attributes tables, albeit being in
partial shape (w/o its own attributes yet).

Currently, tests fail, and it seems there's a memory problem. Disabling
the GC allows a sample script to work, while with GC running it fails with
a "cannot compare" error. The script is currently in code.pln.
Should debug the memory problems and then proceed with making the changes
to the object system, in order to finally have cell_table and table work
with Value keys and not char* keys.